fix(cjs): build cjs for `loadEnv`
