+++ packages/vite/src/node/publicUtils.ts
@@ -11,3 +11,4 @@ export { normalizePath, mergeConfig, mergeAlias } from './utils'
 export { send } from './server/send'
 export { createLogger } from './logger'
 export { searchForWorkspaceRoot } from './server/searchRoot'
+export { loadEnv, resolveEnvPrefix } from './env'
